compute submissions metadata in domain layer
