+++ apps/server/src/modules/task/uc/task.uc.spec.ts
@@ -1,15 +1,17 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { PaginationQuery } from '@shared/controller/dto/pagination.query';
-import { Submission, Task, UserTaskInfo, CourseTaskInfo } from '../entity';
+import { Submission, Task } from '../entity';
 import { SubmissionRepo } from '../repo/submission.repo';
 import { Counted } from '../../../shared/domain';
 import { TaskRepo } from '../repo/task.repo';
 import { TaskUC } from './task.uc';
+import { TaskSubmissionMetadataService } from '../domain/task-submission-metadata.service';
 
 describe('TaskService', () => {
 	let service: TaskUC;
 	let taskRepo: TaskRepo;
 	let submissionRepo: SubmissionRepo;
+	let taskSubmissionMetadata: TaskSubmissionMetadataService;
 
 	beforeEach(async () => {
 		const module: TestingModule = await Test.createTestingModule({
@@ -31,12 +33,19 @@ describe('TaskService', () => {
 						getAllSubmissionsByUser() {},
 					},
 				},
+				{
+					provide: TaskSubmissionMetadataService,
+					useValue: {
+						computeSubmissionMetadata() {},
+					},
+				},
 			],
 		}).compile();
 
 		service = module.get<TaskUC>(TaskUC);
 		submissionRepo = module.get<SubmissionRepo>(SubmissionRepo);
 		taskRepo = module.get<TaskRepo>(TaskRepo);
+		taskSubmissionMetadata = module.get<TaskSubmissionMetadataService>(TaskSubmissionMetadataService);
 	});
 
 	it('should be defined', () => {
@@ -77,9 +86,11 @@ describe('TaskService', () => {
 				.mockImplementation(() => {
 					return Promise.resolve([[], 0] as Counted<Submission[]>);
 				});
-			const computeSubmissionMetadataSpy = jest.spyOn(service, 'computeSubmissionMetadata').mockImplementation(() => {
-				return { submitted: 0, maxSubmissions: 1, graded: 0 };
-			});
+			const computeSubmissionMetadataSpy = jest
+				.spyOn(taskSubmissionMetadata, 'computeSubmissionMetadata')
+				.mockImplementation(() => {
+					return { submitted: 0, maxSubmissions: 1, graded: 0 };
+				});
 
 			const [tasks, total] = await service.findAllOpenByTeacher('abc', new PaginationQuery());
 			expect(total).toEqual(2);
@@ -102,7 +113,7 @@ describe('TaskService', () => {
 					return Promise.resolve([[{ task: task1 }, { task: task2 }], 0] as Counted<Submission[]>);
 				});
 			const computeSubmissionMetadataSpy = jest
-				.spyOn(service, 'computeSubmissionMetadata')
+				.spyOn(taskSubmissionMetadata, 'computeSubmissionMetadata')
 				.mockImplementation((submissions: Submission[]) => {
 					return { submitted: submissions.length, maxSubmissions: 0, graded: 0 };
 				});
@@ -115,79 +126,5 @@ describe('TaskService', () => {
 		});
 
 		// teacher only --> should not needed
-	});
-
-	describe('getTaskSubmissionMetadata', () => {
-		it('should return the number of students that submitted', () => {
-			const task = new Task({ course: new CourseTaskInfo({}) });
-			const testdata = [
-				new Submission({ student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }) }),
-				new Submission({ student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'def' }) }),
-			];
-
-			const result = service.computeSubmissionMetadata(testdata, task);
-			expect(result.submitted).toEqual(2);
-		});
-
-		it('should count submissions by the same student only once', () => {
-			const task = new Task({ course: new CourseTaskInfo({}) });
-			const testdata = [
-				new Submission({ student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }) }),
-				new Submission({ student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }) }),
-			];
-
-			const result = service.computeSubmissionMetadata(testdata, task);
-			expect(result.submitted).toEqual(1);
-		});
-
-		it('should return the maximum number of students that could submit', () => {
-			const task = { name: 'name', course: { students: [{ id: 'abc' }, { id: 'def' }] } } as unknown;
-
-			const result = service.computeSubmissionMetadata([], task as Task);
-			expect(result.maxSubmissions).toEqual(2);
-		});
-
-		it('should return the number of submissions that have been graded', () => {
-			const task = new Task({ course: new CourseTaskInfo({}) });
-			const testdata = [
-				new Submission({
-					student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }),
-					grade: 50,
-				}),
-				new Submission({
-					student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'def' }),
-					gradeComment: 'well done',
-				}),
-				// TODO: add grade file case
-				/* new Submission({
-							student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'def' }),
-							gradeFileIds: [new FileTaskInfo({})],
-						}), */
-				new Submission({
-					student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'def' }),
-				}),
-			];
-
-			const result = service.computeSubmissionMetadata(testdata, task);
-			expect(result.graded).toEqual(2);
-		});
-
-		it('should consider only the newest submission per user for grading', () => {
-			const task = new Task({ course: new CourseTaskInfo({}) });
-			const testdata = [
-				new Submission({
-					student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }),
-					createdAt: new Date(Date.now()),
-				}),
-				new Submission({
-					student: new UserTaskInfo({ firstName: 'firstname', lastName: 'lastname', id: 'abc' }),
-					gradeComment: 'well done',
-					createdAt: new Date(Date.now() - 500),
-				}),
-			];
-
-			const result = service.computeSubmissionMetadata(testdata, task);
-			expect(result.graded).toEqual(1);
-		});
 	});
 });
