+++ apps/server/src/modules/task/uc/task.uc.ts
@@ -6,10 +6,10 @@ import { EntityId, IPagination } from '@shared/domain';
 import { Counted } from '@shared/domain/types';
 import { ICurrentUser } from '../../authentication/interface/jwt-payload';
 import { TaskRepo } from '../repo/task.repo';
-import { Task, Submission, ISubmissionStatus } from '../entity';
 import { SubmissionRepo } from '../repo/submission.repo';
 import { TaskMapper } from '../mapper/task.mapper';
 import { TaskResponse } from '../controller/dto';
+import { TaskSubmissionMetadataService } from '../domain/task-submission-metadata.service';
 
 // TODO: add time filter filter in uc
 @Injectable()
@@ -19,38 +19,11 @@ export class TaskUC {
 		studentDashboard: 'TASK_DASHBOARD_VIEW_V3',
 	};
 
-	constructor(private taskRepo: TaskRepo, private submissionRepo: SubmissionRepo) {}
-
-	// TODO: must move to other location, it is more a private methode but a part that must be tested seperatly
-	computeSubmissionMetadata = (taskSubmissions: Submission[], task: Task): ISubmissionStatus => {
-		const submittedUsers = new Set();
-		const gradedUsers = new Set();
-
-		const sortedSubmissions = [...taskSubmissions].sort((a: Submission, b: Submission) => {
-			if (a.createdAt > b.createdAt) {
-				return 1;
-			}
-			return -1;
-		});
-
-		sortedSubmissions.forEach((submission) => {
-			if (
-				!submittedUsers.has(submission.student.id) &&
-				(submission.grade || submission.gradeComment || submission.gradeFileIds)
-			) {
-				gradedUsers.add(submission.student.id);
-			}
-			submittedUsers.add(submission.student.id);
-		});
-		// TODO: consider coursegroups
-		const studentsInTasksCourse = task.course.students.length;
-
-		return {
-			submitted: submittedUsers.size,
-			maxSubmissions: studentsInTasksCourse,
-			graded: gradedUsers.size,
-		};
-	};
+	constructor(
+		private taskRepo: TaskRepo,
+		private submissionRepo: SubmissionRepo,
+		private taskSubmissionMetadata: TaskSubmissionMetadataService
+	) {}
 
 	// TODO: Combine student and teacher logic if it is possible in next iterations
 	// TODO: Add for students in status -> student has finished, teacher has answered
@@ -74,7 +47,10 @@ export class TaskUC {
 
 		const computedTasks = tasks.map((task) => {
 			const taskSubmissions = submissions.filter((sub) => sub.task === task);
-			return TaskMapper.mapToResponse(task, this.computeSubmissionMetadata(taskSubmissions, task));
+			return TaskMapper.mapToResponse(
+				task,
+				this.taskSubmissionMetadata.computeSubmissionMetadata(taskSubmissions, task)
+			);
 		});
 		return [computedTasks, total];
 	}
